+++ framework/singleton.py
@@ -1,7 +1,7 @@
 # Singleton class - inherit it for only a single object of a given class to exist
+from abc import ABC
 
-
-class Singleton:
+class Singleton(ABC):
 
     def __new__(cls, *args, **kwargs):
         if not hasattr(cls, 'instance'):