+++ views.py
@@ -1,30 +1,28 @@
 # REFERENCE DESIGN
-from framework.response import Response
-from framework.request import Request
-from framework.views import View, BaseView, SimpleView, MinimalView
-from framework.singleton import Singleton
-from framework.urls import Url
-
-# Imports for Contact class
 from datetime import datetime
 import os
+
+from framework.response import Response
+from framework.request import Request
+from framework.views import View, BaseView, MinimalView
+from framework.singleton import Singleton
+from framework.persistence import Persistence
+
 import settings
+from category import Category
+from course import Course
 
 
 # Main page controller  - SINGLETON! Initialize it with menu urls when Framework is initialized!
 class Index(View, Singleton):
-    HTML_TEMPLATE = settings.PATH_APP + "index_body.html"             # Page template
-
-    # Initialize urls property with the passed array, or with empty array if not yet initialized
-    def __init__(self, urls: [Url] = None):
-        if urls is not None:
-            self.urls = urls
-        elif not hasattr(self, 'urls'):
-            self.urls = []
-        #print("urls {}: param: {} self: {}".format(id(self), urls, self.urls))
+    HTML_TEMPLATE = settings.PATH_APP + "index.html"             # Page template
 
     def run(self, request: Request, *args, **kwargs) -> Response:
-        return BaseView(self.HTML_TEMPLATE, title="Главная страница", urls=self.urls).run(request)
+        return BaseView(self.HTML_TEMPLATE,
+                        title="Главная страница",
+                        categories=Persistence.engine(Category).get(),
+                        courses=Persistence.engine(Course).get(),
+                        ).run(request)
 
 
 # Page Controller для формы обратной связи
@@ -37,8 +35,8 @@ class Index(View, Singleton):
 # Метод POST записывает данные формы в папку сообщений в файл с именем в формате YYYY-MM-DDTHH-MM-SS_email.
 class Contact(View):
 
-    HTML_TEMPLATE_FORM = settings.PATH_APP + "contact_body.html"             # Шаблон формы обратной связи
-    HTML_TEMPLATE_SUCCESS = settings.PATH_APP + "contact_success_body.html"  # Шаблон страницы успеха
+    HTML_TEMPLATE_FORM = settings.PATH_APP + "contact.html"             # Шаблон формы обратной связи
+    HTML_TEMPLATE_SUCCESS = settings.PATH_APP + "contact_success.html"  # Шаблон страницы успеха
     PATH_CONTACT_MESSAGES = settings.PATH_APP + "contact_messages/"     # Папка для файлов сообщений
     QPARAM_NAME = "name"                                                # Параметры POST с полями сообщения
     QPARAM_EMAIL = "email"
@@ -70,12 +68,11 @@ class Contact(View):
                     output_file.write(message)
             except FileNotFoundError:
                 return MinimalView(status_code=404,
-                                   body="Папка сообщений не найдена - возможно, отсутствуют права доступа.").run()
+                                   body="Папка сообщений не найдена - возможно, отсутствуют права доступа.")\
+                    .run(request)
 
                                             # Отобразить страницу успеха
             return BaseView(self.HTML_TEMPLATE_SUCCESS, message=message_html, *args, **kwargs).run(request)
 
         else:                                                   # Неизвестный метод - ошибка
-            return MinimalView(status_code=418, body="Неизвестный метод HTTP: {}".format(request.method)).run()
-
-
+            return MinimalView(status_code=418, body="Неизвестный метод HTTP: {}".format(request.method)).run(request)